Improve import logic
Added a --path option, which will append the specified path to
sys.path. If the path starts with a / it is assumed to be a fully
qualified uri. If not then it takes the relative path and tacs that
onto the end of os.getcwd() and then appends that to sys.path.

Once the path is appended, __import__ is called to import the specified
module from the given path.